Fixed network version number in netMessages

+++ Solution/Network/NetMessage.h
@@ -23,15 +23,11 @@ public:
 	uint32_t myTimeStamp;
 	uint32_t mySenderID;
 	uint32_t myTargetID;
-	uint16_t myGameID;
+	uint32_t myGameID;
 	uint8_t myID;
 protected:
-
 	virtual void DoSerialize(StreamType& aStream);
 	virtual void DoDeSerialize(StreamType& aStream);
 	virtual void DoDeSerializeFirst(StreamType& aStream);
-
-
-
 };
